// src/app/core/models/active-user-enabled-data.model.ts
export class ActiveUserEnabledData {
  enableAreas: boolean; // المنطقة
  enableBranches: boolean; // الفرع
  enableDepartments: boolean; // ادارة
  enableGovernments: boolean; // المحافظة
  enableSources: boolean; // مصدر ابلاغ
  enableUniversities: boolean; // الجامعات
  showAreas: boolean;
  showBranches: boolean;
  showDepartments: boolean;
  showGovernments: boolean;
  showSources: boolean;
  showUniversities: boolean;
  showGovernmentEventLevel: boolean;
  showDepartmentEventLevel: boolean;
  showSourceEventLevel: boolean;
  showBranchEventLevel: boolean;
  showUniversityEventLevel: boolean;
  showAreaEventLevel: boolean;
}